Update tentacat and use its ReviewRequests API
We were on a fork of tentacat because it didn't support Github's API for review requests yet, but it
has since been updated.  This switches us to the latest tentacat off the main repo, and updates our
app to use the correct API call.  The custom header in the config file is necessary because the
reviews stuff is still in a pre-release state.